# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Movie API
  version: 1.0.0
servers:
- url: https://localhost:8080/api/v1
...
paths:
...
  /bookings:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Booking'
                type: array
          description: OK
      summary: List all bookings
      tags:
      - Bookings
      x-accepts: application/json
      x-tags:
      - tag: Bookings
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Booking'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
          description: Created
      summary: Create a new booking
      tags:
      - Bookings
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Bookings
  /bookings/{id}:
    delete:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "204":
          description: No Content
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_bookings__id__get_404_response'
          description: The specified booking was not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_movies__id__put_500_response'
          description: An internal server error occurred
      summary: Cancel a booking
      tags:
      - Bookings
      x-accepts: application/json
      x-tags:
      - tag: Bookings
    get:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_bookings__id__get_404_response'
          description: The specified booking was not found
      summary: Get details of a specific booking
      tags:
      - Bookings
      x-accepts: application/json
      x-tags:
      - tag: Bookings
    put:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Booking'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_bookings__id__get_404_response'
          description: The specified booking was not found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_bookings__id__put_409_response'
          description: A booking with the same ID already exists
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_bookings__id__put_422_response'
          description: There are not enough tickets available for the requested showtime
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_movies__id__put_500_response'
          description: An internal server error occurred
      summary: "Update details of a specific booking (e.g., add/remove tickets)"
      tags:
      - Bookings
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Bookings
...
components:
...
  schemas:
...
    Booking:
      example:
        totalAmount: 5.962133916683182
        customerId: 6
        showtimeId: 1
        bookingDate: 2000-01-23T04:56:07.000+00:00
        bookingId: 0
      properties:
        bookingId:
          description: Unique identifier for the booking
          format: int64
          type: integer
        customerId:
          description: ID of the customer who made the booking
          format: int64
          type: integer
        showtimeId:
          description: ID of the showtime associated with the booking
          format: int64
          type: integer
        bookingDate:
          description: Date and time the booking was made
          format: date-time
          type: string
        totalAmount:
          description: Total amount paid for the booking
          format: double
          type: number
      type: object
...
    _movies__id__put_500_response:
      properties:
        message:
          description: Generic error message
          type: string
      type: object
...
    _bookings__id__get_404_response:
      properties:
        message:
          description: Error message indicating booking not found
          type: string
      type: object
    _bookings__id__put_409_response:
      properties:
        message:
          description: Error message indicating duplicate booking ID
          type: string
      type: object
    _bookings__id__put_422_response:
      properties:
        message:
          description: Error message indicating insufficient tickets
          type: string
      type: object
...